Move the `storage.azure` and `storage.s3` packages into the `storage.providers` namespace
