+++ RELEASE.md
@@ -1,0 +1,3 @@
+RELEASE_TYPE: minor
+
+Move the `storage.azure` and `storage.s3` packages into the `storage.providers` namespace for consistency.
