+++ storage/src/main/scala/weco/storage/listing/azure/AzureListing.scala
@@ -1,8 +1,7 @@
 package weco.storage.listing.azure
 
-import weco.storage.azure.AzureBlobLocationPrefix
 import weco.storage.listing.Listing
-import weco.storage.azure.AzureBlobLocationPrefix
 import weco.storage.listing.Listing
+import weco.storage.providers.azure.AzureBlobLocationPrefix
 
 trait AzureListing[Result] extends Listing[AzureBlobLocationPrefix, Result]
