+++ storage/src/main/scala/weco/storage/listing/s3/S3Listing.scala
@@ -1,7 +1,7 @@
 package weco.storage.listing.s3
 
 import weco.storage.listing.Listing
-import weco.storage.s3.S3ObjectLocationPrefix
+import weco.storage.providers.s3.S3ObjectLocationPrefix
 
 trait S3Listing[ListingResult]
     extends Listing[S3ObjectLocationPrefix, ListingResult]
