+++ storage/src/main/scala/weco/storage/store/dynamo/DynamoVersionedHybridStore.scala
@@ -1,8 +1,8 @@
 package weco.storage.store.dynamo
 
-import weco.storage.s3.S3ObjectLocation
+import weco.storage.providers.s3.S3ObjectLocation
 import weco.storage.store.VersionedHybridStore
-import weco.storage.s3.S3ObjectLocation
+import weco.storage.providers.s3.S3ObjectLocation
 
 class DynamoVersionedHybridStore[Id, V, T](
   store: DynamoHybridStoreWithMaxima[Id, V, T])(implicit N: Numeric[V])
