+++ storage/src/main/scala/weco/storage/store/s3/S3TypedStore.scala
@@ -1,7 +1,7 @@
 package weco.storage.store.s3
 
 import software.amazon.awssdk.services.s3.S3Client
-import weco.storage.s3.S3ObjectLocation
+import weco.storage.providers.s3.S3ObjectLocation
 import weco.storage.store.TypedStore
 import weco.storage.streaming.Codec
