+++ storage/src/test/scala/weco/storage/fixtures/S3Fixtures.scala
@@ -6,7 +6,10 @@ import io.circe.{Decoder, Json}
 import org.scalatest.concurrent.{Eventually, IntegrationPatience}
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.{Assertion, EitherValues}
-import software.amazon.awssdk.auth.credentials.{AwsBasicCredentials, StaticCredentialsProvider}
+import software.amazon.awssdk.auth.credentials.{
+  AwsBasicCredentials,
+  StaticCredentialsProvider
+}
 import software.amazon.awssdk.core.sync.RequestBody
 import software.amazon.awssdk.services.s3.model._
 import software.amazon.awssdk.services.s3.presigner.S3Presigner
@@ -14,7 +17,7 @@ import software.amazon.awssdk.services.s3.{S3Client, S3Configuration}
 import weco.fixtures._
 import weco.json.JsonUtil._
 import weco.storage.generators.{S3ObjectLocationGenerators, StreamGenerators}
-import weco.storage.s3.{S3Config, S3ObjectLocation}
+import weco.storage.providers.s3.{S3Config, S3ObjectLocation}
 import weco.storage.streaming.Codec._
 import weco.storage.streaming.InputStreamWithLength
