+++ storage/src/test/scala/weco/storage/listing/s3/S3ListingFixtures.scala
@@ -5,7 +5,7 @@ import weco.fixtures.TestWith
 import weco.storage.fixtures.S3Fixtures
 import weco.storage.fixtures.S3Fixtures.Bucket
 import weco.storage.listing.fixtures.ListingFixtures
-import weco.storage.s3.{S3ObjectLocation, S3ObjectLocationPrefix}
+import weco.storage.providers.s3.{S3ObjectLocation, S3ObjectLocationPrefix}
 
 trait S3ListingFixtures[ListingResult]
     extends S3Fixtures
