+++ storage/src/test/scala/weco/storage/listing/s3/S3ListingTestCases.scala
@@ -5,7 +5,7 @@ import software.amazon.awssdk.services.s3.model.NoSuchBucketException
 import weco.fixtures.TestWith
 import weco.storage.fixtures.S3Fixtures.Bucket
 import weco.storage.listing.ListingTestCases
-import weco.storage.s3.{S3ObjectLocation, S3ObjectLocationPrefix}
+import weco.storage.providers.s3.{S3ObjectLocation, S3ObjectLocationPrefix}
 
 trait S3ListingTestCases[ListingResult]
     extends ListingTestCases[
