+++ storage/src/test/scala/weco/storage/store/dynamo/DynamoHybridStoreTest.scala
@@ -6,7 +6,7 @@ import weco.storage.{StoreReadError, StoreWriteError, Version}
 import weco.storage.fixtures.DynamoFixtures.Table
 import weco.storage.fixtures.S3Fixtures.Bucket
 import weco.storage.generators.Record
-import weco.storage.s3.S3ObjectLocation
+import weco.storage.providers.s3.S3ObjectLocation
 
 import scala.language.higherKinds
