+++ storage/src/test/scala/weco/storage/store/fixtures/NamespaceFixtures.scala
@@ -3,7 +3,7 @@ package weco.storage.store.fixtures
 import weco.fixtures.{RandomGenerators, TestWith}
 import weco.storage.fixtures.S3Fixtures
 import weco.storage.fixtures.S3Fixtures.Bucket
-import weco.storage.s3.S3ObjectLocation
+import weco.storage.providers.s3.S3ObjectLocation
 
 trait NamespaceFixtures[Ident, Namespace] {
   def withNamespace[R](testWith: TestWith[Namespace, R]): R
